Move the data drift metrics into their own module
